feat(frontend): island like landforming
- `components/explorer.vue` is replaced with `components/explorer2.vue`.
- Each subcluster is rendered as an island. Papers in a subcluster are
  rendered as dots on the island if zoomed enough. A pre-rendered data
  is utilized to skip an expensive steps for arranging clusters, papers
  and probability contours.
